Drafts merged w/ events, fixed editor tracking, event editing issue
- The person who edited an event is now show on the changelog
- Fixed issue with updating event generating Embed within a new thread that doesn't have a transaction
- Fixed updating an event with no note and not updating the note
- End date added to reminder triggers
- Events now get status changed to completed when the event time comes.
- Drafts have been merged into events. Now with a status to represent the state of the event

+++ src/main/java/com/th3hero/eventbot/jobs/DeletedEventCleanupJob.java
@@ -37,6 +37,10 @@ public class DeletedEventCleanupJob implements Job {
         final EventJpa eventJpa = eventRepository.findById(eventId)
             .orElseThrow(() -> new EntityNotFoundException("Attempted to clean up deleted event that does not exist. Event id: %d".formatted(eventId)));
 
+        if (!eventJpa.getStatus().equals(EventJpa.EventStatus.DELETED)) {
+            throw new IllegalStateException("Attempted to cleanup an event that was no longer deleted. Cleanup trigger should have been removed. Event id: %d".formatted(eventId));
+        }
+
         final ConfigJpa config = configService.getConfigJpa();
 
         eventRepository.delete(eventJpa);
